Refactor the sidebar into separate components
