+++ src/page/home/sidebar/ChatSwitcherView.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class ChatSwitcherView extends React.Component {
+    render() {
+        return (
+            <View>
+                <Text>Chat Switcher</Text>
+            </View>
+        );
+    }
+}
+
+export default ChatSwitcherView;
